scrumTools/src/test/java/ch/paru/scrumTools/capacity/release/factories/ReleaseDataFactoryTest.java
package ch.paru.scrumTools.capacity.release.factories;

import static org.junit.Assert.assertNotNull;

import org.easymock.EasyMockSupport;
import org.junit.Test;

import ch.paru.scrumTools.capacity.release.data.ReleaseData;
import ch.paru.scrumTools.capacity.shared.factories.TeamFactory;
import ch.paru.scrumTools.capacity.shared.factories.TeamMemberFactory;

public class ReleaseDataFactoryTest {

	private static final EasyMockSupport MOCKS = new EasyMockSupport();

	@Test
	public void testInstance() {
		MOCKS.resetAll();

		MOCKS.replayAll();
		ReleaseDataFactory factory = new ReleaseDataFactory();
		ReleaseData result = factory.createData(new TeamFactory(), new TeamMemberFactory());

		MOCKS.verifyAll();
		assertNotNull(result);
	}

}
